empire/bulk-queue/requirements.txt
redis[hiredis]==5.2.1
anthropic==0.42.0
httpx==0.28.1
python-dotenv==1.0.1
